gilang

+++ application/modules/investortype/controllers/RequestController.php
@@ -9,6 +9,7 @@ class Investortype_RequestController extends MyIndo_Controller_Action
 	protected $_error_code;
 	protected $_error_message;
 	protected $_success;
+	protected $_data;
 	
 	public function init()
 	{
@@ -27,42 +28,37 @@ class Investortype_RequestController extends MyIndo_Controller_Action
 		$this->_error_code = 0;
 		$this->_error_message = '';
 		$this->_success = true;
+		
+		$this->_data = array(
+				'data' => array(
+						'items' => array(),
+						'totalCount' => 0
+						)
+				);
 	}
 	
 	public function createAction()
 	{
-		$data = array(
-			'data' => array()
-		);
-		
-		if($this->getRequest()->isPost() && $this->getRequest()->isXmlHttpRequest()) {
+		if($this->isPost() && $this->isAjax()) {
 			try {
-				
 				if(!$this->_model->isExistByKey('INVESTOR_TYPE', $this->_posts['INVESTOR_TYPE'])) {
 					//Insert Data :
 					$this->_model->insert(array(
 							'INVESTOR_TYPE'=> $this->_posts['INVESTOR_TYPE'],
 							'CREATED_DATE' => date('Y-m-d H:i:s')
 							));
-				
 				} else {
-					$this->_error_code = 201;
-					$this->_error_message = MyIndo_Tools_Error::getErrorMessage($this->_error_code);
-					$this->_success = false;
+					$this->error(201);
 				}
-				
 			}catch(Exception $e) {
 				$this->_error_code = $e->getCode();
 				$this->_error_message = $e->getMessage();
 				$this->_success = false;
 			}
 		} else {
-			$this->_error_code = 901;
-			$this->_error_message = MyIndo_Tools_Error::getErrorMessage($this->_error_code);
-			$this->_success = false;
+			$this->error(901);
 		}
-		
-		MyIndo_Tools_Return::JSON($data, $this->_error_code, $this->_error_message, $this->_success);
+		$this->json();
 	}
 	
 	public function readAction()
@@ -114,10 +110,6 @@ class Investortype_RequestController extends MyIndo_Controller_Action
 	
 	public function updateAction()
 	{
-		$data = array(
-				'data' => array()
-		);
-		
 		try {
 			$posts = $this->getRequest()->getRawBody();
 			$posts = Zend_Json::decode($posts);
@@ -131,25 +123,26 @@ class Investortype_RequestController extends MyIndo_Controller_Action
 			$this->_error_message = $e->getMessage();
 			$this->_success = false;
 		}
-		
-		MyIndo_Tools_Return::JSON($data, $this->_error_code, $this->_error_message, $this->_success);
+		$this->json();
 	}
 	
 	public function destroyAction()
 	{
-		$data = array(
-				'data' => array()
-				);
 		try {
- 			$this->_model->delete(
- 					$this->_model->getAdapter()->quoteInto(
- 							'INVESTOR_TYPE_ID = ?', $this->_posts['INVESTOR_TYPE_ID']
- 							));
+			$modelInvestors = new Application_Model_Investors();
+			if(!$modelInvestors->isExistByKey('INVESTOR_TYPE_ID', $this->_posts['INVESTOR_TYPE_ID'])) {
+	 			$this->_model->delete(
+	 					$this->_model->getAdapter()->quoteInto(
+	 							'INVESTOR_TYPE_ID = ?', $this->_posts['INVESTOR_TYPE_ID']
+	 							));
+			} else {
+				$this->error(202);
+			}
 		}catch(Exception $e) {
 			$this->_error_code = $e->getCode();
 			$this->_error_message = $e->getMessage();
 			$this->_success = false;
 		}
-		MyIndo_Tools_Return::JSON($data, $this->_error_code, $this->_error_message, $this->_success);
+		$this->json();
 	}
 }
